Changes that support 1.3.3
## Added

**Devices and labels are now unlimited!**

Important: you must increase the size of the card to see more labels/devices. Just adding them to the card configuration / YAML is not enough, the card hides them if rows/column minimums aren't met.

Cap limits are in place for the default size of the card (12 columns, 3 rows): 2 Grid, 2 Battery, 4 PV labels, 8 Devices.

Note: There are of course theoretical limits due to the size of the cards which I cannot really avoid without making the UI tiny. In testing I am seeing 8 devices and PV labels for every 12 columns used, in theory grid and battery labels should be unlimited if you keep growing the card height. 

Another Note: On small screens (phones), the width of wide cards is squeezed small - in this case a max of 8 PV labels or devices will be shown in the UI. 

Reminder: Card resizing is [managed using sections](https://github.com/pacemaker82/Compact-Power-Card#managing-the-size-of-the-card).

**Your mileage may vary with this change. It can lead to overlapping labels and other undesirable effects as there are so many possible scenarios that I can't really control.**

## Other Changes

- `curve_factor` updated to support 0-5 range.
- `curved_lines` is now deprecated. Use `curve_factor: 0` for same effect. 
- Refactored all geometry across render and update to remove duplication.